agent/requirements.txt
requests==2.31.0
PyYAML==6.0
kubernetes==25.3.0
PyGithub==2.1.2
python-dotenv==1.0.0

